Moved all of the condense functions into a single file and deleted the originals.
